+++ Makefile.am
@@ -2,11 +2,8 @@ AM_CFLAGS = -g -Wall -D_GNU_SOURCE -I$(srcdir)/include
 
 bin_PROGRAMS = \
 	simple/fi_info \
-	simple/fi_pingpong \
-	simple/fi_write_lat \
-	simple/fi_write_bw \
-	simple/fi_read_lat \
-	simple/fi_read_bw \
+	simple/fi_msg_pingpong \
+	simple/fi_msg_rma \
 	simple/fi_ud_pingpong \
 	unit/fi_eq_test \
 	unit/fi_av_test \
@@ -16,24 +13,12 @@ simple_fi_info_SOURCES = \
 	simple/info.c \
 	common/shared.c
 
-simple_fi_pingpong_SOURCES = \
-	simple/pingpong.c \
+simple_fi_msg_pingpong_SOURCES = \
+	simple/msg_pingpong.c \
 	common/shared.c
 
-simple_fi_write_lat_SOURCES = \
-	simple/write_lat.c \
-	common/shared.c
-
-simple_fi_write_bw_SOURCES = \
-	simple/write_bw.c \
-	common/shared.c
-
-simple_fi_read_lat_SOURCES = \
-	simple/read_lat.c \
-	common/shared.c
-
-simple_fi_read_bw_SOURCES = \
-	simple/read_bw.c \
+simple_fi_msg_rma_SOURCES = \
+	simple/msg_rma.c \
 	common/shared.c
 
 simple_fi_ud_pingpong_SOURCES = \
