+++ simple/msg_pingpong.c
@@ -36,10 +36,9 @@
 #include <unistd.h>
 
 #include <rdma/fabric.h>
-#include <rdma/fi_endpoint.h>
-#include <rdma/fi_rma.h>
-#include <rdma/fi_cm.h>
 #include <rdma/fi_errno.h>
+#include <rdma/fi_endpoint.h>
+#include <rdma/fi_cm.h>
 #include <shared.h>
 
 static int custom;
@@ -47,12 +46,10 @@ static int size_option;
 static int iterations = 1000;
 static int transfer_size = 1000;
 static int max_credits = 128;
-static int warmup_iters = 128;
+static int credits = 128;
 static char test_name[10] = "custom";
 static struct timespec start, end;
 static void *buf;
-static void *rem_buf;
-static uint64_t rem_key;
 static size_t buffer_size;
 
 static struct fi_info hints;
@@ -69,105 +66,98 @@ static struct fid_eq *cmeq;
 static struct fid_cq *rcq, *scq;
 static struct fid_mr *mr;
 
-static void show_perf(void)
+static int send_xfer(int size)
 {
-	int64_t usec = get_elapsed(&start, &end, MICRO);
-	long long bytes = (long long) iterations * transfer_size * 2;
-
-	perf_str(test_name, transfer_size, iterations, bytes, usec);
-}
-
-static void init_test(int size)
-{
-	char sstr[32];
-
-	size_str(sstr, size);
-	snprintf(test_name, sizeof test_name, "%s_bw", sstr);
-	transfer_size = size;
-	iterations = size_to_count(transfer_size);
-}
-
-static int send_msg(int size)
-{
+	struct fi_cq_entry comp;
 	int ret;
 
+	while (!credits) {
+		ret = fi_cq_read(scq, &comp, 1);
+		if (ret > 0) {
+			goto post;
+		} else if (ret < 0) {
+			if (ret == -FI_EAVAIL) {
+				cq_readerr(scq, "scq");
+			} else {
+				printf("Completion queue read %d (%s)\n", ret, fi_strerror(-ret));
+			}
+			return ret;
+		}
+	}
+
+	credits--;
+post:
 	ret = fi_send(ep, buf, (size_t) size, fi_mr_desc(mr), NULL);
-	if (ret) {
-		fprintf(stderr, "fi_send %d (%s)\n", ret, fi_strerror(-ret));
-		return ret;
-	}
-
-	return wait_for_completion(scq, 1);
-}
-
-static int post_recv()
-{
-	int ret;
-
-	ret = fi_recv(ep, buf, buffer_size, fi_mr_desc(mr), NULL);
-	if (ret) {
-		fprintf(stderr, "fi_recv %d (%s)\n", ret, fi_strerror(-ret));
-		return ret;
-	}
+	if (ret)
+		printf("fi_send %d (%s)\n", ret, fi_strerror(-ret));
 
 	return ret;
 }
 
-static int write_data(size_t size)
+static int recv_xfer(int size)
+{
+	struct fi_cq_entry comp;
+	int ret;
+
+	do {
+		ret = fi_cq_read(rcq, &comp, 1);
+		if (ret < 0) {
+			if (ret == -FI_EAVAIL) {
+				cq_readerr(rcq, "rcq");
+			} else {
+				printf("Completion queue read %d (%s)\n", ret, fi_strerror(-ret));
+			}
+			return ret;
+		}
+	} while (!ret);
+
+	ret = fi_recv(ep, buf, buffer_size, fi_mr_desc(mr), buf);
+	if (ret)
+		printf("fi_recv %d (%s)\n", ret, fi_strerror(-ret));
+
+	return ret;
+}
+
+static int sync_test(void)
 {
 	int ret;
 
-	ret = fi_write(ep, buf, size, fi_mr_desc(mr),  
-		       (uint64_t)rem_buf, rem_key, NULL);
+	ret = wait_for_completion(scq, max_credits - credits);
 	if (ret) {
-		fprintf(stderr, "fi_write %d (%s)\n", ret, fi_strerror(-ret));
 		return ret;
 	}
-	return 0;
-}
+	credits = max_credits;
 
-static int warmup(int iters)
-{
-	int ret, i;
-	for (i = 0; i<iters; i++) {
-		if ((ret = write_data(16)) < 0)
-			return ret;
+	ret = dst_addr ? send_xfer(16) : recv_xfer(16);
+	if (ret) {
+		return ret;
 	}
 
-	if ((ret = wait_for_completion(scq, iters)) < 0)
-		return ret;
-
-	return 0;
+	return dst_addr ? recv_xfer(16) : send_xfer(16);
 }
 
 static int run_test(void)
 {
-	int ret, i, oust;
+	int ret, i;
 
-	ret = warmup(MIN(warmup_iters, max_credits));
+	ret = sync_test();
 	if (ret)
 		goto out;
 
 	clock_gettime(CLOCK_MONOTONIC, &start);
-	for (i = 0, oust =0; i < iterations; i++, oust++) {
-		ret = write_data(transfer_size);
+	for (i = 0; i < iterations; i++) {
+		ret = dst_addr ? send_xfer(transfer_size) :
+				 recv_xfer(transfer_size);
 		if (ret)
 			goto out;
 
-		if (oust == max_credits) {
-			ret = wait_for_completion(scq, oust);
-			if (ret)
-				goto out;
-			oust = 0;
-		}
+		ret = dst_addr ? recv_xfer(transfer_size) :
+				 send_xfer(transfer_size);
+		if (ret)
+			goto out;
 	}
-
-	ret = wait_for_completion(scq, oust);
-	if (ret)
-		goto out;
-
 	clock_gettime(CLOCK_MONOTONIC, &end);
-	show_perf();
+	show_perf(test_name, transfer_size, iterations, &start, &end, 2);
 	ret = 0;
 
 out:
@@ -188,7 +178,7 @@ static int alloc_cm_res(void)
 	cm_attr.wait_obj = FI_WAIT_FD;
 	ret = fi_eq_open(fab, &cm_attr, &cmeq, NULL);
 	if (ret)
-		fprintf(stderr, "fi_eq_open cm %s\n", fi_strerror(-ret));
+		printf("fi_eq_open cm %s\n", fi_strerror(-ret));
 
 	return ret;
 }
@@ -207,7 +197,7 @@ static int alloc_ep_res(struct fi_info *fi)
 	int ret;
 
 	buffer_size = !custom ? test_size[TEST_CNT - 1].size : transfer_size;
-	buf = malloc(MAX(buffer_size, sizeof(uint64_t)));
+	buf = malloc(buffer_size);
 	if (!buf) {
 		perror("malloc");
 		return -1;
@@ -219,20 +209,19 @@ static int alloc_ep_res(struct fi_info *fi)
 	cq_attr.size = max_credits << 1;
 	ret = fi_cq_open(dom, &cq_attr, &scq, NULL);
 	if (ret) {
-		fprintf(stderr, "fi_eq_open send comp %s\n", fi_strerror(-ret));
+		printf("fi_eq_open send comp %s\n", fi_strerror(-ret));
 		goto err1;
 	}
 
 	ret = fi_cq_open(dom, &cq_attr, &rcq, NULL);
 	if (ret) {
-		fprintf(stderr, "fi_eq_open recv comp %s\n", fi_strerror(-ret));
+		printf("fi_eq_open recv comp %s\n", fi_strerror(-ret));
 		goto err2;
 	}
-	
-	ret = fi_mr_reg(dom, buf, MAX(buffer_size, sizeof(uint64_t)), 
-			FI_REMOTE_WRITE, 0, 0, 0, &mr, NULL);
+
+	ret = fi_mr_reg(dom, buf, buffer_size, 0, 0, 0, 0, &mr, NULL);
 	if (ret) {
-		fprintf(stderr, "fi_mr_reg %s\n", fi_strerror(-ret));
+		printf("fi_mr_reg %s\n", fi_strerror(-ret));
 		goto err3;
 	}
 
@@ -265,7 +254,7 @@ static int bind_ep_res(void)
 		return ret;
 	}
 
-	ret = fi_bind(&ep->fid, &scq->fid, FI_SEND|FI_WRITE);
+	ret = fi_bind(&ep->fid, &scq->fid, FI_SEND);
 	if (ret) {
 		printf("fi_bind %s\n", fi_strerror(-ret));
 		return ret;
@@ -281,6 +270,10 @@ static int bind_ep_res(void)
 	if (ret)
 		return ret;
 
+	ret = fi_recv(ep, buf, buffer_size, fi_mr_desc(mr), buf);
+	if (ret)
+		printf("fi_recv %d (%s)\n", ret, fi_strerror(-ret));
+
 	return ret;
 }
 
@@ -291,19 +284,19 @@ static int server_listen(void)
 
 	ret = fi_getinfo(FI_VERSION(1, 0), src_addr, port, FI_SOURCE, &hints, &fi);
 	if (ret) {
-		fprintf(stderr, "fi_getinfo %s\n", strerror(-ret));
+		printf("fi_getinfo %s\n", strerror(-ret));
 		return ret;
 	}
 
 	ret = fi_fabric(fi->fabric_attr, &fab, NULL);
 	if (ret) {
-		fprintf(stderr, "fi_fabric %s\n", fi_strerror(-ret));
+		printf("fi_fabric %s\n", fi_strerror(-ret));
 		goto err0;
 	}
 
 	ret = fi_pendpoint(fab, fi, &pep, NULL);
 	if (ret) {
-		fprintf(stderr, "fi_endpoint %s\n", fi_strerror(-ret));
+		printf("fi_endpoint %s\n", fi_strerror(-ret));
 		goto err1;
 	}
 
@@ -319,7 +312,7 @@ static int server_listen(void)
 
 	ret = fi_listen(pep);
 	if (ret) {
-		fprintf(stderr, "fi_listen %s\n", fi_strerror(-ret));
+		printf("fi_listen %s\n", fi_strerror(-ret));
 		goto err3;
 	}
 
@@ -346,12 +339,12 @@ static int server_connect(void)
 
 	rd = fi_eq_sread(cmeq, &event, &entry, sizeof entry, -1, 0);
 	if (rd != sizeof entry) {
-		fprintf(stderr, "fi_eq_sread %zd %s\n", rd, fi_strerror((int) -rd));
+		printf("fi_eq_sread %zd %s\n", rd, fi_strerror((int) -rd));
 		return (int) rd;
 	}
 
 	if (event != FI_CONNREQ) {
-		fprintf(stderr, "Unexpected CM event %d\n", event);
+		printf("Unexpected CM event %d\n", event);
 		ret = -FI_EOTHER;
 		goto err1;
 	}
@@ -359,14 +352,13 @@ static int server_connect(void)
 	info = entry.info;
 	ret = fi_domain(fab, info, &dom, NULL);
 	if (ret) {
-		fprintf(stderr, "fi_fdomain %s\n", fi_strerror(-ret));
+		printf("fi_fdomain %s\n", fi_strerror(-ret));
 		goto err1;
 	}
 
-
 	ret = fi_endpoint(dom, info, &ep, NULL);
 	if (ret) {
-		fprintf(stderr, "fi_endpoint for req %s\n", fi_strerror(-ret));
+		printf("fi_endpoint for req %s\n", fi_strerror(-ret));
 		goto err1;
 	}
 
@@ -380,35 +372,34 @@ static int server_connect(void)
 
 	ret = fi_accept(ep, NULL, 0);
 	if (ret) {
-		fprintf(stderr, "fi_accept %s\n", fi_strerror(-ret));
+		printf("fi_accept %s\n", fi_strerror(-ret));
 		goto err3;
 	}
 
 	rd = fi_eq_sread(cmeq, &event, &entry, sizeof entry, -1, 0);
- 	if (rd != sizeof entry) {
-		fprintf(stderr, "fi_eq_sread %zd %s\n", rd, fi_strerror((int) -rd));
+	if (rd != sizeof entry) {
+		printf("fi_eq_sread %zd %s\n", rd, fi_strerror((int) -rd));
 		goto err3;
- 	}
+	}
 
 	if (event != FI_COMPLETE || entry.fid != &ep->fid) {
- 		fprintf(stderr, "Unexpected CM event %d fid %p (ep %p)\n",
+		printf("Unexpected CM event %d fid %p (ep %p)\n",
 			event, entry.fid, ep);
- 		ret = -FI_EOTHER;
- 		goto err3;
- 	}
- 
- 	fi_freeinfo(info);
- 	return 0;
+		ret = -FI_EOTHER;
+		goto err3;
+	}
+
+	fi_freeinfo(info);
+	return 0;
 
 err3:
 	free_ep_res();
 err2:
 	fi_close(&ep->fid);
 err1:
-
- 	fi_reject(pep, info->connreq, NULL, 0);
- 	fi_freeinfo(info);
- 	return ret;
+	fi_reject(pep, info->connreq, NULL, 0);
+	fi_freeinfo(info);
+	return ret;
 }
 
 static int client_connect(void)
@@ -423,31 +414,31 @@ static int client_connect(void)
 		ret = getaddr(src_addr, NULL, (struct sockaddr **) &hints.src_addr,
 			      (socklen_t *) &hints.src_addrlen);
 		if (ret)
-			fprintf(stderr, "source address error %s\n", gai_strerror(ret));
+			printf("source address error %s\n", gai_strerror(ret));
 	}
 
 	ret = fi_getinfo(FI_VERSION(1, 0), dst_addr, port, 0, &hints, &fi);
 	if (ret) {
-		fprintf(stderr, "fi_getinfo %s\n", strerror(-ret));
+		printf("fi_getinfo %s\n", strerror(-ret));
 		goto err0;
 	}
 
 	ret = fi_fabric(fi->fabric_attr, &fab, NULL);
 	if (ret) {
-		fprintf(stderr, "fi_fabric %s\n", fi_strerror(-ret));
+		printf("fi_fabric %s\n", fi_strerror(-ret));
 		goto err1;
 	}
 
- 	ret = fi_domain(fab, fi, &dom, NULL);
+	ret = fi_domain(fab, fi, &dom, NULL);
 	if (ret) {
-		fprintf(stderr, "fi_fdomain %s %s\n", fi_strerror(-ret),
+		printf("fi_fdomain %s %s\n", fi_strerror(-ret),
 			fi->domain_attr->name);
 		goto err2;
 	}
 
 	ret = fi_endpoint(dom, fi, &ep, NULL);
 	if (ret) {
-		fprintf(stderr, "fi_endpoint %s\n", fi_strerror(-ret));
+		printf("fi_endpoint %s\n", fi_strerror(-ret));
 		goto err3;
 	}
 
@@ -461,22 +452,22 @@ static int client_connect(void)
 
 	ret = fi_connect(ep, fi->dest_addr, NULL, 0);
 	if (ret) {
-		fprintf(stderr, "fi_connect %s\n", fi_strerror(-ret));
+		printf("fi_connect %s\n", fi_strerror(-ret));
 		goto err5;
 	}
 
- 	rd = fi_eq_sread(cmeq, &event, &entry, sizeof entry, -1, 0);
+	rd = fi_eq_sread(cmeq, &event, &entry, sizeof entry, -1, 0);
 	if (rd != sizeof entry) {
-		fprintf(stderr, "fi_eq_condread %zd %s\n", rd, fi_strerror((int) -rd));
+		printf("fi_eq_condread %zd %s\n", rd, fi_strerror((int) -rd));
 		return (int) rd;
 	}
 
- 	if (event != FI_COMPLETE || entry.fid != &ep->fid) {
- 		fprintf(stderr, "Unexpected CM event %d fid %p (ep %p)\n",
- 			event, entry.fid, ep);
- 		ret = -FI_EOTHER;
- 		goto err1;
- 	}
+	if (event != FI_COMPLETE || entry.fid != &ep->fid) {
+		printf("Unexpected CM event %d fid %p (ep %p)\n",
+			event, entry.fid, ep);
+		ret = -FI_EOTHER;
+		goto err1;
+	}
 
 	if (hints.src_addr)
 		free(hints.src_addr);
@@ -499,28 +490,6 @@ err0:
 	return ret;
 }
 
-static void exchange_params(void)
-{
-	*((uint64_t *) buf) = (uint64_t) buf;
-	*(((uint64_t *) buf + 1)) = fi_mr_key(mr);
-	post_recv();
-	send_msg(sizeof(uint64_t *) * 2);
-	wait_for_completion(rcq, 1);
-
-	rem_buf = (void *) (*((uint64_t *) buf));
-	rem_key = (uint64_t) (*(((uint64_t *) buf + 1)));
-}
-
-static void synchronize(void)
-{
-	if (dst_addr) {
-		post_recv();
-		wait_for_completion(rcq, 1);
-	} else {
-		send_msg(sizeof(uint64_t *));
-	}
-}
-
 static int run(void)
 {
 	int i, ret = 0;
@@ -531,30 +500,30 @@ static int run(void)
 			return ret;
 	}
 
-	printf("%-10s%-8s%-8s%-8s%8s %10s%13s\n",
-	       "name", "bytes", "xfers", "total", "time", "Gb/sec", "usec/xfer");
+	print_test_hdr();
 
 	ret = dst_addr ? client_connect() : server_connect();
-	if (ret)
+	if (ret) {
 		return ret;
-
-	exchange_params();
+	}
 
 	if (!custom) {
 		for (i = 0; i < TEST_CNT; i++) {
 			if (test_size[i].option > size_option)
 				continue;
-			init_test(test_size[i].size);
-			ret = run_test();
-			if(ret)
-				goto out;
+			init_test(test_size[i].size, test_name, &transfer_size, &iterations);
+			run_test();
 		}
 	} else {
 		ret = run_test();
 	}
-	synchronize();
 
-out:
+	ret = wait_for_completion(scq, max_credits - credits);
+	if (ret) {
+		return ret;
+	}
+	credits = max_credits;
+
 	fi_shutdown(ep, 0);
 	fi_close(&ep->fid);
 	free_ep_res();
@@ -569,7 +538,7 @@ int main(int argc, char **argv)
 {
 	int op, ret;
 
-	while ((op = getopt(argc, argv, "d:n:p:s:I:w:S:")) != -1) {
+	while ((op = getopt(argc, argv, "d:n:p:s:I:S:")) != -1) {
 		switch (op) {
 		case 'd':
 			dst_addr = optarg;
@@ -587,9 +556,6 @@ int main(int argc, char **argv)
 			custom = 1;
 			iterations = atoi(optarg);
 			break;
-		case 'w':
-			warmup_iters = atoi(optarg);
-			break;
 		case 'S':
 			if (!strncasecmp("all", optarg, 3)) {
 				size_option = 1;
@@ -599,14 +565,13 @@ int main(int argc, char **argv)
 			}
 			break;
 		default:
-			fprintf(stderr, "usage: %s\n", argv[0]);
-			fprintf(stderr, "\t[-d destination_address]\n");
-			fprintf(stderr, "\t[-n domain_name]\n");
-			fprintf(stderr, "\t[-p port_number]\n");
-			fprintf(stderr, "\t[-s source_address]\n");
-			fprintf(stderr, "\t[-I iterations]\n");
-			fprintf(stderr, "\t[-w warmup iterations]\n");
-			fprintf(stderr, "\t[-S transfer_size or 'all']\n");
+			printf("usage: %s\n", argv[0]);
+			printf("\t[-d destination_address]\n");
+			printf("\t[-n domain_name]\n");
+			printf("\t[-p port_number]\n");
+			printf("\t[-s source_address]\n");
+			printf("\t[-I iterations]\n");
+			printf("\t[-S transfer_size or 'all']\n");
 			exit(1);
 		}
 	}
@@ -614,7 +579,7 @@ int main(int argc, char **argv)
 	hints.domain_attr = &domain_hints;
 	hints.ep_attr = &ep_hints;
 	hints.ep_type = FI_EP_MSG;
-	hints.caps = FI_MSG | FI_RMA;
+	hints.caps = FI_MSG;
 	hints.mode = FI_LOCAL_MR | FI_PROV_MR_KEY;
 	hints.addr_format = FI_SOCKADDR;
