+++ simple/msg_rma.c
@@ -33,14 +33,15 @@
 #include <getopt.h>
 #include <time.h>
 #include <netdb.h>
-#include <unistd.h>
 
 #include <rdma/fabric.h>
-#include <rdma/fi_errno.h>
 #include <rdma/fi_endpoint.h>
+#include <rdma/fi_rma.h>
 #include <rdma/fi_cm.h>
+#include <rdma/fi_errno.h>
 #include <shared.h>
 
+static uint64_t op_type = FI_REMOTE_WRITE;
 static int custom;
 static int size_option;
 static int iterations = 1000;
@@ -51,6 +52,7 @@ static char test_name[10] = "custom";
 static struct timespec start, end;
 static void *buf;
 static size_t buffer_size;
+struct fi_rma_iov local, remote;
 
 static struct fi_info hints;
 static struct fi_domain_attr domain_hints;
@@ -66,22 +68,17 @@ static struct fid_eq *cmeq;
 static struct fid_cq *rcq, *scq;
 static struct fid_mr *mr;
 
-static void show_perf(void)
+void usage(char *name)
 {
-	int64_t usec = get_elapsed(&start, &end, MICRO);
-	long long bytes = (long long) iterations * transfer_size * 2;
-
-	perf_str(test_name, transfer_size, iterations, bytes, usec);
-}
-
-static void init_test(int size)
-{
-	char sstr[32];
-
-	size_str(sstr, size);
-	snprintf(test_name, sizeof test_name, "%s_lat", sstr);
-	transfer_size = size;
-	iterations = size_to_count(transfer_size);
+	fprintf(stderr, "usage: %s\n", name);
+	fprintf(stderr, "\t[-d destination_address]\n");
+	fprintf(stderr, "\t[-n domain_name]\n");
+	fprintf(stderr, "\t[-p port_number]\n");
+	fprintf(stderr, "\t[-s source_address]\n");
+	fprintf(stderr, "\t[-I iterations]\n");
+	fprintf(stderr, "\t[-o read|write] (default: write)\n");
+	fprintf(stderr, "\t[-S transfer_size or 'all']\n");
+	exit(1);
 }
 
 static int send_xfer(int size)
@@ -136,6 +133,33 @@ static int recv_xfer(int size)
 	return ret;
 }
 
+static int read_data(size_t size)
+{
+	int ret;
+
+	ret = fi_read(ep, buf, size, fi_mr_desc(mr), 
+		       remote.addr, remote.key, NULL);
+	if (ret) {
+		fprintf(stderr, "fi_read %d (%s)\n", ret, fi_strerror(-ret));
+		return ret;
+	}
+
+	return 0;
+}
+
+static int write_data(size_t size)
+{
+	int ret;
+
+	ret = fi_write(ep, buf, size, fi_mr_desc(mr),  
+		       remote.addr, remote.key, NULL);
+	if (ret) {
+		fprintf(stderr, "fi_write %d (%s)\n", ret, fi_strerror(-ret));
+		return ret;
+	}
+	return 0;
+}
+
 static int sync_test(void)
 {
 	int ret;
@@ -158,24 +182,23 @@ static int run_test(void)
 {
 	int ret, i;
 
-	ret = sync_test();
-	if (ret)
-		goto out;
+	sync_test();
 
 	clock_gettime(CLOCK_MONOTONIC, &start);
 	for (i = 0; i < iterations; i++) {
-		ret = dst_addr ? send_xfer(transfer_size) :
-				 recv_xfer(transfer_size);
+		if (op_type == FI_REMOTE_WRITE) {
+			ret = write_data(transfer_size);
+		} else {
+			ret = read_data(transfer_size); 
+		}
 		if (ret)
 			goto out;
-
-		ret = dst_addr ? recv_xfer(transfer_size) :
-				 send_xfer(transfer_size);
+		ret = wait_for_completion(scq, 1);
 		if (ret)
 			goto out;
 	}
 	clock_gettime(CLOCK_MONOTONIC, &end);
-	show_perf();
+	show_perf(test_name, transfer_size, iterations, &start, &end, 1);
 	ret = 0;
 
 out:
@@ -196,7 +219,7 @@ static int alloc_cm_res(void)
 	cm_attr.wait_obj = FI_WAIT_FD;
 	ret = fi_eq_open(fab, &cm_attr, &cmeq, NULL);
 	if (ret)
-		printf("fi_eq_open cm %s\n", fi_strerror(-ret));
+		fprintf(stderr, "fi_eq_open cm %s\n", fi_strerror(-ret));
 
 	return ret;
 }
@@ -215,7 +238,7 @@ static int alloc_ep_res(struct fi_info *fi)
 	int ret;
 
 	buffer_size = !custom ? test_size[TEST_CNT - 1].size : transfer_size;
-	buf = malloc(buffer_size);
+	buf = malloc(MAX(buffer_size, sizeof(uint64_t)));
 	if (!buf) {
 		perror("malloc");
 		return -1;
@@ -227,19 +250,20 @@ static int alloc_ep_res(struct fi_info *fi)
 	cq_attr.size = max_credits << 1;
 	ret = fi_cq_open(dom, &cq_attr, &scq, NULL);
 	if (ret) {
-		printf("fi_eq_open send comp %s\n", fi_strerror(-ret));
+		fprintf(stderr, "fi_eq_open send comp %s\n", fi_strerror(-ret));
 		goto err1;
 	}
 
 	ret = fi_cq_open(dom, &cq_attr, &rcq, NULL);
 	if (ret) {
-		printf("fi_eq_open recv comp %s\n", fi_strerror(-ret));
+		fprintf(stderr, "fi_eq_open recv comp %s\n", fi_strerror(-ret));
 		goto err2;
 	}
-
-	ret = fi_mr_reg(dom, buf, buffer_size, 0, 0, 0, 0, &mr, NULL);
+	
+	ret = fi_mr_reg(dom, buf, MAX(buffer_size, sizeof(uint64_t)), 
+			op_type, 0, 0, 0, &mr, NULL);
 	if (ret) {
-		printf("fi_mr_reg %s\n", fi_strerror(-ret));
+		fprintf(stderr, "fi_mr_reg %s\n", fi_strerror(-ret));
 		goto err3;
 	}
 
@@ -302,19 +326,19 @@ static int server_listen(void)
 
 	ret = fi_getinfo(FI_VERSION(1, 0), src_addr, port, FI_SOURCE, &hints, &fi);
 	if (ret) {
-		printf("fi_getinfo %s\n", strerror(-ret));
+		fprintf(stderr, "fi_getinfo %s\n", strerror(-ret));
 		return ret;
 	}
 
 	ret = fi_fabric(fi->fabric_attr, &fab, NULL);
 	if (ret) {
-		printf("fi_fabric %s\n", fi_strerror(-ret));
+		fprintf(stderr, "fi_fabric %s\n", fi_strerror(-ret));
 		goto err0;
 	}
 
 	ret = fi_pendpoint(fab, fi, &pep, NULL);
 	if (ret) {
-		printf("fi_endpoint %s\n", fi_strerror(-ret));
+		fprintf(stderr, "fi_endpoint %s\n", fi_strerror(-ret));
 		goto err1;
 	}
 
@@ -330,7 +354,7 @@ static int server_listen(void)
 
 	ret = fi_listen(pep);
 	if (ret) {
-		printf("fi_listen %s\n", fi_strerror(-ret));
+		fprintf(stderr, "fi_listen %s\n", fi_strerror(-ret));
 		goto err3;
 	}
 
@@ -357,12 +381,12 @@ static int server_connect(void)
 
 	rd = fi_eq_sread(cmeq, &event, &entry, sizeof entry, -1, 0);
 	if (rd != sizeof entry) {
-		printf("fi_eq_sread %zd %s\n", rd, fi_strerror((int) -rd));
+		fprintf(stderr, "fi_eq_sread %zd %s\n", rd, fi_strerror((int) -rd));
 		return (int) rd;
 	}
 
 	if (event != FI_CONNREQ) {
-		printf("Unexpected CM event %d\n", event);
+		fprintf(stderr, "Unexpected CM event %d\n", event);
 		ret = -FI_EOTHER;
 		goto err1;
 	}
@@ -370,13 +394,14 @@ static int server_connect(void)
 	info = entry.info;
 	ret = fi_domain(fab, info, &dom, NULL);
 	if (ret) {
-		printf("fi_fdomain %s\n", fi_strerror(-ret));
+		fprintf(stderr, "fi_fdomain %s\n", fi_strerror(-ret));
 		goto err1;
 	}
 
+
 	ret = fi_endpoint(dom, info, &ep, NULL);
 	if (ret) {
-		printf("fi_endpoint for req %s\n", fi_strerror(-ret));
+		fprintf(stderr, "fi_endpoint for req %s\n", fi_strerror(-ret));
 		goto err1;
 	}
 
@@ -390,34 +415,35 @@ static int server_connect(void)
 
 	ret = fi_accept(ep, NULL, 0);
 	if (ret) {
-		printf("fi_accept %s\n", fi_strerror(-ret));
+		fprintf(stderr, "fi_accept %s\n", fi_strerror(-ret));
 		goto err3;
 	}
 
 	rd = fi_eq_sread(cmeq, &event, &entry, sizeof entry, -1, 0);
-	if (rd != sizeof entry) {
-		printf("fi_eq_sread %zd %s\n", rd, fi_strerror((int) -rd));
+ 	if (rd != sizeof entry) {
+		fprintf(stderr, "fi_eq_sread %zd %s\n", rd, fi_strerror((int) -rd));
 		goto err3;
-	}
+ 	}
 
 	if (event != FI_COMPLETE || entry.fid != &ep->fid) {
-		printf("Unexpected CM event %d fid %p (ep %p)\n",
+ 		fprintf(stderr, "Unexpected CM event %d fid %p (ep %p)\n",
 			event, entry.fid, ep);
-		ret = -FI_EOTHER;
-		goto err3;
-	}
-
-	fi_freeinfo(info);
-	return 0;
+ 		ret = -FI_EOTHER;
+ 		goto err3;
+ 	}
+ 
+ 	fi_freeinfo(info);
+ 	return 0;
 
 err3:
 	free_ep_res();
 err2:
 	fi_close(&ep->fid);
 err1:
-	fi_reject(pep, info->connreq, NULL, 0);
-	fi_freeinfo(info);
-	return ret;
+
+ 	fi_reject(pep, info->connreq, NULL, 0);
+ 	fi_freeinfo(info);
+ 	return ret;
 }
 
 static int client_connect(void)
@@ -432,31 +458,31 @@ static int client_connect(void)
 		ret = getaddr(src_addr, NULL, (struct sockaddr **) &hints.src_addr,
 			      (socklen_t *) &hints.src_addrlen);
 		if (ret)
-			printf("source address error %s\n", gai_strerror(ret));
+			fprintf(stderr, "source address error %s\n", gai_strerror(ret));
 	}
 
 	ret = fi_getinfo(FI_VERSION(1, 0), dst_addr, port, 0, &hints, &fi);
 	if (ret) {
-		printf("fi_getinfo %s\n", strerror(-ret));
+		fprintf(stderr, "fi_getinfo %s\n", strerror(-ret));
 		goto err0;
 	}
 
 	ret = fi_fabric(fi->fabric_attr, &fab, NULL);
 	if (ret) {
-		printf("fi_fabric %s\n", fi_strerror(-ret));
+		fprintf(stderr, "fi_fabric %s\n", fi_strerror(-ret));
 		goto err1;
 	}
 
-	ret = fi_domain(fab, fi, &dom, NULL);
+ 	ret = fi_domain(fab, fi, &dom, NULL);
 	if (ret) {
-		printf("fi_fdomain %s %s\n", fi_strerror(-ret),
+		fprintf(stderr, "fi_fdomain %s %s\n", fi_strerror(-ret),
 			fi->domain_attr->name);
 		goto err2;
 	}
 
 	ret = fi_endpoint(dom, fi, &ep, NULL);
 	if (ret) {
-		printf("fi_endpoint %s\n", fi_strerror(-ret));
+		fprintf(stderr, "fi_endpoint %s\n", fi_strerror(-ret));
 		goto err3;
 	}
 
@@ -470,22 +496,22 @@ static int client_connect(void)
 
 	ret = fi_connect(ep, fi->dest_addr, NULL, 0);
 	if (ret) {
-		printf("fi_connect %s\n", fi_strerror(-ret));
+		fprintf(stderr, "fi_connect %s\n", fi_strerror(-ret));
 		goto err5;
 	}
 
-	rd = fi_eq_sread(cmeq, &event, &entry, sizeof entry, -1, 0);
+ 	rd = fi_eq_sread(cmeq, &event, &entry, sizeof entry, -1, 0);
 	if (rd != sizeof entry) {
-		printf("fi_eq_condread %zd %s\n", rd, fi_strerror((int) -rd));
+		fprintf(stderr, "fi_eq_condread %zd %s\n", rd, fi_strerror((int) -rd));
 		return (int) rd;
 	}
 
-	if (event != FI_COMPLETE || entry.fid != &ep->fid) {
-		printf("Unexpected CM event %d fid %p (ep %p)\n",
-			event, entry.fid, ep);
-		ret = -FI_EOTHER;
-		goto err1;
-	}
+ 	if (event != FI_COMPLETE || entry.fid != &ep->fid) {
+ 		fprintf(stderr, "Unexpected CM event %d fid %p (ep %p)\n",
+ 			event, entry.fid, ep);
+ 		ret = -FI_EOTHER;
+ 		goto err1;
+ 	}
 
 	if (hints.src_addr)
 		free(hints.src_addr);
@@ -508,6 +534,26 @@ err0:
 	return ret;
 }
 
+static int exchange_addr_key(void)
+{
+	local.addr = (uint64_t)buf;
+	local.key = fi_mr_key(mr);
+
+	if (dst_addr) {
+		*(struct fi_rma_iov *)buf = local;
+		send_xfer(sizeof local);
+		recv_xfer(sizeof remote);
+		remote = *(struct fi_rma_iov *)buf;
+	} else {
+		recv_xfer(sizeof remote);
+		remote = *(struct fi_rma_iov *)buf;
+		*(struct fi_rma_iov *)buf = local;
+		send_xfer(sizeof local);
+	}
+
+	return 0;
+}
+
 static int run(void)
 {
 	int i, ret = 0;
@@ -518,31 +564,32 @@ static int run(void)
 			return ret;
 	}
 
-	printf("%-10s%-8s%-8s%-8s%8s %10s%13s\n",
-		"name", "bytes", "iters", "total", "time", "Gb/sec", "usec/xfer");
+	print_test_hdr();
 
 	ret = dst_addr ? client_connect() : server_connect();
-	if (ret) {
+	if (ret)
 		return ret;
-	}
+
+	ret = exchange_addr_key();
+	if (ret)
+		return ret;
 
 	if (!custom) {
 		for (i = 0; i < TEST_CNT; i++) {
 			if (test_size[i].option > size_option)
 				continue;
-			init_test(test_size[i].size);
-			run_test();
+			init_test(test_size[i].size, test_name, &transfer_size, &iterations);
+			ret = run_test();
+			if(ret)
+				goto out;
 		}
 	} else {
 		ret = run_test();
 	}
 
-	ret = wait_for_completion(scq, max_credits - credits);
-	if (ret) {
-		return ret;
-	}
-	credits = max_credits;
+	sync_test();
 
+out:
 	fi_shutdown(ep, 0);
 	fi_close(&ep->fid);
 	free_ep_res();
@@ -557,7 +604,7 @@ int main(int argc, char **argv)
 {
 	int op, ret;
 
-	while ((op = getopt(argc, argv, "d:n:p:s:I:S:")) != -1) {
+	while ((op = getopt(argc, argv, "d:n:p:s:I:o:S:")) != -1) {
 		switch (op) {
 		case 'd':
 			dst_addr = optarg;
@@ -575,6 +622,14 @@ int main(int argc, char **argv)
 			custom = 1;
 			iterations = atoi(optarg);
 			break;
+		case 'o':
+			if (!strcmp(optarg, "read"))
+				op_type = FI_REMOTE_READ;
+			else if (!strcmp(optarg, "write"))
+				op_type = FI_REMOTE_WRITE;
+			else
+				usage(argv[0]);
+			break;
 		case 'S':
 			if (!strncasecmp("all", optarg, 3)) {
 				size_option = 1;
@@ -584,21 +639,14 @@ int main(int argc, char **argv)
 			}
 			break;
 		default:
-			printf("usage: %s\n", argv[0]);
-			printf("\t[-d destination_address]\n");
-			printf("\t[-n domain_name]\n");
-			printf("\t[-p port_number]\n");
-			printf("\t[-s source_address]\n");
-			printf("\t[-I iterations]\n");
-			printf("\t[-S transfer_size or 'all']\n");
-			exit(1);
+			usage(argv[0]);
 		}
 	}
 
 	hints.domain_attr = &domain_hints;
 	hints.ep_attr = &ep_hints;
 	hints.ep_type = FI_EP_MSG;
-	hints.caps = FI_MSG;
+	hints.caps = FI_MSG | FI_RMA;
 	hints.mode = FI_LOCAL_MR | FI_PROV_MR_KEY;
 	hints.addr_format = FI_SOCKADDR;
